refactor: factor out computronCounter
