+++ packages/cosmic-swingset/src/launch-chain.js
@@ -42,15 +42,11 @@ import {
   makeSlogCallbacks,
 } from './kernel-stats.js';
 
-import {
-  BeansPerBlockComputeLimit,
-  BeansPerVatCreation,
-  BeansPerXsnapComputron,
-} from './sim-params.js';
 import { parseParams } from './params.js';
 import { makeQueue, makeQueueStorageMock } from './helpers/make-queue.js';
 import { exportStorage } from './export-storage.js';
 import { parseLocatedJson } from './helpers/json.js';
+import { computronCounter } from './computron-counter.js';
 
 /** @import { Mailbox, RunPolicy, SwingSetConfig } from '@agoric/swingset-vat' */
 /** @import { KVStore } from './helpers/bufferedStorage.js' */
@@ -250,67 +246,6 @@ export async function buildSwingset(
  *   totalBeans(): bigint;
  * }} ChainRunPolicy
  */
-
-/**
- * @typedef {object} BeansPerUnit
- * @property {bigint} blockComputeLimit
- * @property {bigint} vatCreation
- * @property {bigint} xsnapComputron
- */
-
-/**
- * @param {BeansPerUnit} beansPerUnit
- * @param {boolean} [ignoreBlockLimit]
- * @returns {ChainRunPolicy}
- */
-function computronCounter(
-  {
-    [BeansPerBlockComputeLimit]: blockComputeLimit,
-    [BeansPerVatCreation]: vatCreation,
-    [BeansPerXsnapComputron]: xsnapComputron,
-  },
-  ignoreBlockLimit = false,
-) {
-  assert.typeof(blockComputeLimit, 'bigint');
-  assert.typeof(vatCreation, 'bigint');
-  assert.typeof(xsnapComputron, 'bigint');
-  let totalBeans = 0n;
-  const shouldRun = () => ignoreBlockLimit || totalBeans < blockComputeLimit;
-  const remainingBeans = () =>
-    ignoreBlockLimit ? undefined : blockComputeLimit - totalBeans;
-
-  const policy = harden({
-    vatCreated() {
-      totalBeans += vatCreation;
-      return shouldRun();
-    },
-    crankComplete(details = {}) {
-      assert.typeof(details, 'object');
-      if (details.computrons) {
-        assert.typeof(details.computrons, 'bigint');
-
-        // TODO: xsnapComputron should not be assumed here.
-        // Instead, SwingSet should describe the computron model it uses.
-        totalBeans += details.computrons * xsnapComputron;
-      }
-      return shouldRun();
-    },
-    crankFailed() {
-      const failedComputrons = 1000000n; // who knows, 1M is as good as anything
-      totalBeans += failedComputrons * xsnapComputron;
-      return shouldRun();
-    },
-    emptyCrank() {
-      return shouldRun();
-    },
-    shouldRun,
-    remainingBeans,
-    totalBeans() {
-      return totalBeans;
-    },
-  });
-  return policy;
-}
 
 export async function launch({
   actionQueueStorage,
